requirements.txt
Django==1.10.5
appdirs==1.4.0
django-widget-tweaks==1.4.1
packaging==16.8
pyparsing==2.1.10
six==1.10.0
django-taggit==0.22.0
